fix: address  feedback for signal robustness and test coverage - Wrapped all post_save publishers (snapshot, chapter, event) in   transaction.on_commit() to prevent phantom Redis messages during rollbacks - Fixed falsy dictionary parsing in signals so empty strings aren't converted to None - Updated 0073 migration to use PositiveBigIntegerField for GenericForeignKey   object_id to prevent overflow with large integer keys - Added missing days_remaining payload assertion in deadline reminder test - Fixed chapter and event signal tests to properly cover the meaningful-update   publish path instead of just the no-change skip path - Added transaction.on_commit mock decorators to test classes to pass Pytest strict   database-access rules

+++ backend/apps/owasp/migrations/0073_notification_subscription.py
@@ -51,7 +51,7 @@ class Migration(migrations.Migration):
                         auto_created=True, primary_key=True, serialize=False, verbose_name="ID"
                     ),
                 ),
-                ("object_id", models.PositiveIntegerField()),
+                ("object_id", models.PositiveBigIntegerField()),
                 ("created_at", models.DateTimeField(auto_now_add=True)),
                 (
                     "content_type",
